Replaced .piskel images with .png versions and more image stuff
Took every .piskel and created a new one, allowing it to be opened and
used. Also created a folder so they don’t have to be hanging around in
a class and instead can have their own place until their used.

So glad there are no more looooong file extensions.