Created bad method, trying to work on good
Is it possible to get the index of an item in an array as an int?
If so, I’ll do that to determine a) the owner of the square and b)
other values using the index. If anyone knows how please tell me!

Otherwise, I could use my inefficient commented-out method…