Made Buttons Responsive, Along with Other Stuff
The buttons now are responsible and tell you when you’ve clicked! Yay!
Otherwise, all I’ve done is create a start for the actual gameplay
mechanism and tried to set the screen color to green (with no success
yet.)

Hopefully I’ll be reminded to read the comment that I’ve left at the
bottom:

TO DO LIST:
- Read other to do list

Bye for now