Added Colored Borders, Created Info Page
The squares now all have a colored border. It allows it to blend in, so
when I import the .pngs onto the buttons it will show as a board
without white space in between the squares. This will just make it look
better, but hey, it’s worth it.

Also, I created an information page that can be used by the player (not
yet usable) in a .txt file that will be able to show them more
information about commands and whatnot to make the gameplay easier.

There will probably be another push in a few minutes.