Parrot AI Added
Technically. In this update I’ve only added in the code for the Parrot
AI player (which is actually extremely helpful because 60% of their
moves are just adding in infrastructure and armies) but the internal
methods being used for claiming are still terrible and don’t work. I
fixed one bug as well but otherwise didn’t do much.