Dodo Bird AI Completed
This is also along the whole “kind of” finished thing. Each method for
an AI involves the “aiClaim()” method, which currently has no code and
is a complete failure. Currently, it just serves as a placeholder for
“AI will claim here,” but will have added functionality in the future.
Otherwise, I made no other changes.

Just kidding - I changed every reference to “Eclectus” or “Eclectus
Parrot” to “Dodo.”