Checks and balance update 1
The first check and balance was added - now, for the player (and the player only, the AI still needs this advantage), the amount of infrastructure gained is multiplied by the amount of squares. This ensures that it's impossible to win an economic victory by only working on your economy and nothing on claiming and your military. The game is coming into place and will probably be finished by the end of the week.